setup netlify
