+++ docs/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  rules: {
+    'react/prop-types': 'off',
+    'react/react-in-jsx-scope': 'off',
+  },
+};
