--- docs/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "extends": "next",
-  "rules": {
-    "react/react-in-jsx-scope": "off"
-  }
-}
