+++ docs/package.json
@@ -4,9 +4,10 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "NODE_ENV=production next build",
+    "build": "cross-env NODE_ENV=production next build",
     "preview": "serve ./export",
-    "lint": "next lint"
+    "lint": "next lint",
+    "typescript": "tsc --noEmit -p ."
   },
   "dependencies": {
     "@base_ui/react": "^1.0.0-alpha.3",
