+++ docs/src/theme.ts
@@ -1,6 +1,3 @@
-import {} from 'react';
-import type { SxProp } from '@pigment-css/react';
-
 export const theme = {
   gray: {
     surface: {
@@ -71,17 +68,3 @@ export const theme = {
 };
 
 export type Theme = typeof theme;
-
-declare module '@pigment-css/react/theme' {
-  export interface ThemeArgs {
-    theme: Theme & {
-      vars: Theme;
-    };
-  }
-}
-
-declare module 'react' {
-  interface Attributes {
-    sx?: SxProp | undefined;
-  }
-}
